RavenDB-26452 Split tests into separate files, use client API, fix history tests
Split monolithic AiAgentForkConversation.cs into 12 files with shared
base class. Use client API (store.AI.*, session) instead of
database.DocumentsStorage for fork/snapshot/purge/message verification.
History tests (D1-D4, B4) now use ChatTrimming.Truncate to create actual
history documents. Assert exact message counts, non-empty collections,
and specific values throughout.
